Moved VML into oldie-extension
