+++ js/Extensions/Oldie/VMLAxis3D.js
@@ -9,7 +9,7 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
-import U from '../Utilities.js';
+import U from '../../Core/Utilities.js';
 var addEvent = U.addEvent;
 /* eslint-disable valid-jsdoc */
 var VMLAxis3DAdditions = /** @class */ (function () {
